class FilesLoad broken to Android and Java implementations - FileDownloadAndroid and FileDownload
